Fix early initialization bug in AbstractMemberProposal
In the previous commit that changed that file I did a change which is
reverted by this commit. I thought that `new { ... } with X` is
equivalent to `new X { ... }`, which is not the case. The former uses
early initialization in order to set values in the trait `X`, whereas
the latter just overrides the values.

Fix #1002642