Allow users to specify the spacing around block imports
A new preference option is added to the Scala > Editor preference page,
which allows users to configure if they want to see spaces around a
block import. The default configuration is that the spaces are added,
because this is the behavior of scalariform (whose behavior is hardcoded
and can't be changed).

Fix #1001692