Move AbstractToggleHandler away from PropertyChangeListenerProxy
The `PropertyChangeListenerProxy` class is not well designed. It has to
use a `WeakReference` to keep a reference to a listener in order to
avoid memory leak when the proxy is no longer needed and the listener
can be removed. The only problem is that I don't know if the weak
reference is of any use at all in this case. Since the proxy unregisters
itself only in case a property change event happens, it may never get
unregistered. My debugging sessions showed that the weak references are
never cleared even if there is clearly no use for them anymore. In order
to avoid any memory leaks and confusion about the implementation, I
think it is best to move completely away from this class.