Save only a position instead of a compiler symbol in implicit annotations.
Compiler symbols should not be kept inside UI elements or other long-lived objects. Symbols may retain in memory a full Scala compiler instance, including all the other symbols and types, leading to memory leaks. In this commit I removed the reference to the implicit symbol, instead keeping only the position. In case the user wants to hyperlink, we re-compute the symbol under the given position, and produce a hyperlink.

Fix #1002679