Improve class file check for mixed Scala/Java projects
In case we have a JAR file that belongs to a project that contains both
Scala and Java source code we need a more intelligent check to find out
if the entire JAR should be checked for class files that belong to Scala
source code.

So far, the first found classfile was responsible for specifying if the
JAR file should be scanned further. If the classfile belongs to a Java
source file, the scanning is aborted, which meant that in a mixed
Scala/Java project, no Scala classfiles can be detected.

In such a case we simply do another check, which finds out if the name
of the JAR file contains a Scala version. This way we can be sure that
it was created from Scala source code because it is unlikely that Java
artifacts contain the Scala version in its artifact name. Once we found
that out, we simply ignore the first classfile if it belongs to Java
source code. Instead we continue with the next classfile until we found
one that belongs to Scala source code.

This is inefficient in case a JAR file contains a Scala version name but
does not contain any Scala sources but this is a very unlikely use case.
For now it allows us to handle mixed Scala/Java projects.

Fixes #1002687