Switch back to client-all-2-11
client-all depends on 2.10 and therefore can't be used in Scala IDE. It
would be better to not hardcode the name in the build but I'm too lazy
for now to make it configurable.

Also, the manifest file is removed from the index and put in gitignore
because it can change.