Do not cache the `MarkOccurrenceIndex`
Caching the `MarkOccurrenceIndex` is problematic, since this quickly leads
to an index, that is referencing an AST that is not identical to the most
recent one from the presentation compiler. Since equality for symbols is
implemented as reference equality, this is a problem even if the trees are
semantically equivalent, resulting in erratic behaviour as described in
ticket #1002701.

I've experimented with a solution in the refactoring library, that performs
semantic equality comparisons on symbols, to work around this problem. This
worked as far as the trees where indeed semantically equivalent. In some cases
however, there where semantic differences in the trees, even though the source
code did not change, like positions being available in one one tree,
but absent in the other. Maybe this could be sorted out by digging deeper, but
in order to keep it simple, I'd rather not go this route unless not caching
actually results in noticeable performance problems.

Fix #1002701