Do not log missing cases in outline to error log
We could log it to info instead but it just pollutes the log. Since we
can't fix these cases anyway because we don't log the original source
code (we could log the sources but that would pollute the logs even more
and it is not likely that users will report back this stuff) it is
better to just not log anything. If we find a case that is not working,
we can still write a test for it and fix it.

Fix #1002710