Replace sbt libraries with zinc libraries
With this change, the IDE no longer depends directly on sbt but only on
zinc. The most important advantage is that we no longer need to build
zinc by ourselves - instead we can rely on the prebuilt zinc artifacts.
We still need to build compiler-bridge (former compiler-interface)
because it depends on scalac and therefore is binary incompatible
between minor versions but since the infrastructure to build it already
exists the only change that was needed is to add different artifacts.