Do now access compiler reference in PresentationReporter
This led to a NPE because the compiler reference can't be initialized
early enough. The reporter needs to be fully initialized in the ctor of
Global but Global is the compiler and therefore we can't fully
initialize the reporter.

Since the compiler references is still needed in InteractiveReporter, we
pass the name separately and hope that functionality of
InteractiveReporter is not used before the compiler is fully
initialized.

Fix #1002733