Automatically restart PC in order to avoid false positive errors
Because the PC ofter shows false positives the idea is to automatically
restart the PC, which would get rid of the false positives. After a
restart all of the invalid state in the PC is gone. A better fix would
be to not get invalid state in the PC but I don't understand the PC good
enough to provide such a fix.

The PC is automatically restart in the following situations:

- After a build. This happens every time the build is run.
- After Eclipse gained focus. This happens once Eclipse lost focus
  and gained it afterwards.
- After an editor gained focus. This happens when a different editor
  is selected.

Since restarting the PC is relatively fast and happens asynchronously in
the background, this shouldn't block users in any way.

Fix #1002725