Do not automatically restart PC on builds when errors exist
Restarting the PC after a build when the build found compilation errors
or warnings doesn't seem to be very useful. The purpose of the build is
it to find errors, therefore it is very likely that a build contains
errors. Since restarting the PC is only helpful to get rid of false
positive errors it only seems to be a waste of resources to restart the
PC after a build.
If build automatically is enabled, it is even more likely that a build
contains errors because building happens more often. In this case a
restart is completely useless.