Remove annotations that belong to a removed range of an editor
When text is removed from an editor the annotations that belong to the
removed range need to be removed too because they are no longer valid.

This is especially useful functionality when the user fixes an error. In
such a case, the builder did already create an annotation previously
that needs to be removed. With this change, it can happen that errors
are removed which were not yet fixed. In most cases, the PC will find
them once again and show them to users. In rare cases, the builder needs
to be run to find all problems. This is not optimal behavior but
probably the best what we can get in Scala because we can't run the
builder on every source change, it is simly to slow.

Fix #1002726